remote_storage: unify azure and S3 tests
